requirements.txt
django==4.1.10
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-redis==5.3.0
django-cacheops==7.0.1
celery==5.3.4
django-celery-results==2.5.1
flower==2.0.1
freezegun==1.2.2
boto3==1.28.51
fabric2==3.2.2
requests==2.31.0
flake8==6.1.0
psycopg2==2.9.8
pydantic==2.4.2
pydantic_core==2.10.1
gunicorn==21.2.0
